Fix flake: Allow sending message in integration test even if no transport defined
