+++ integrationTests/PersonRepositoryTest.php
@@ -16,22 +16,6 @@ use Symfony\Component\Uid\Uuid;
 
 class PersonRepositoryTest extends IntegrationTest
 {
-    /**
-     * Keeping a copy of the original state of the person repo in memory to allow restoring after test is finished
-     * to avoid interference with later tests
-     */
-    private PersonRepository $originalPersonRepository;
-
-    public function setUp(): void
-    {
-        $this->originalPersonRepository = $this->getService(PersonRepository::class);
-    }
-
-    public function tearDown(): void
-    {
-        $this->getWriteableContainer()->set(PersonRepository::class, $this->originalPersonRepository);
-    }
-
     public function testItDoesNotFindPersonWhoDoesNotExist(): void
     {
         $sut = $this->getService(PersonRepository::class);
